+++ src/main/java/com/clipboarder/clipboarder/entity/dto/response/UploadResultDTO.java
@@ -1,4 +1,4 @@
-package com.clipboarder.clipboarder.entity.dto;
+package com.clipboarder.clipboarder.entity.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
